Scripts - iSnobal - Improve COD merge script.
Add command return code check after each step to avoid removing the
original SMRF files before successfully merging them.

+++ scripts/iSnobal/CDO_smrf.sh
@@ -1,6 +1,6 @@
 #!/usr/bin/env bash
 
-# Merge inidividual SMRF solar outputs into one netCDF file
+# Merge individual SMRF solar outputs into one netCDF file
 
 CDO_MATH="\
 ir_in = veg_ir_beam + veg_ir_diffuse;\
@@ -13,11 +13,26 @@ sum_snow = ir_snow + vis_snow;\
 # Merge SMRF outputs
 cdo merge veg_*.nc albedo_*.nc cloud_factor.nc smrf_solar_out.nc
 
+if [ $? != 0 ]; then
+  printf "ERROR: Could not merge SMRF files"
+  exit 1
+fi
+
 # Calculate components
 cdo -P 4 expr,"${CDO_MATH}" smrf_solar_out.nc smrf_solar_sum.nc
 
+if [ $? != 0 ]; then
+  printf "ERROR: Could not calculate SMRF components"
+  exit 1
+fi
+
 # Merge SMRF inputs and outputs
 cdo -O -P 4 -z zip_4 merge smrf_solar_* smrf_solar.nc
+
+if [ $? != 0 ]; then
+  printf "ERROR: Could not merge final files"
+  exit 1
+fi
 
 # Cleanup
 rm veg_*.nc albedo_*.nc cloud_factor.nc smrf_solar_out.nc smrf_solar_sum.nc
